+++ WebShop.BusinessLogic/Interfaces/IUser.cs
@@ -2,6 +2,7 @@
 using WebShop.Domain.User.Auth;
 using WebShop.Domain.User.Registration;
 using WebShop.Domain.User.Modify;
+using WebShop.Domain.User.Delivery;
 
 namespace WebShop.BusinessLogic.Interfaces
 {
@@ -13,7 +14,18 @@ namespace WebShop.BusinessLogic.Interfaces
         bool ChangePasswordInDB(ChangePasswordClass pass);
         int GetUserIdBySessionKey(string sessionKey);
         bool IsSessionValid(string key);
-        //UserRegistrationResponse UserRegistrationAction(UserLoginData data);
         UserInfo GetUserInfoById(int id);
+
+
+
+        //====================Delivery=========================
+
+
+        bool AddDeliveryAddress(DeliveryLoc address);
+        bool EditDeliveryAddress(DeliveryLoc address);
+        bool DeleteDeliveryAddress(int id);
+        DeliveryLoc GetDeliveryAddressById(int id);
+
+        //=============================================
     }
 }
